+++ post-trade/src/test/java/com/multiassetoms/posttrade/infrastructure/InMemoryTradeRepositoryTest.java
@@ -64,6 +64,8 @@ class InMemoryTradeRepositoryTest {
                 "005930",
                 OrderSide.BUY,
                 new BigDecimal("10"),
+                new BigDecimal("55000.0000000000"),
+                new BigDecimal("550000"),
                 TradeStatus.CAPTURED,
                 capturedAt,
                 null,